re-implement bio with TS
